fix(appwrite): make sync field TYPES identical across primary, secondary, and code
--verify only checked field existence, not type, so type drift went
undetected and caused "invalid type" push errors (e.g. shift_notes.createdAt
was string on primary but integer everywhere else).

Changes:
- unified_appwrite_setup.js:
  * verify() now also compares each attribute's actual Appwrite type
    against SCHEMA and reports/count mismatches (exit non-zero).
  * new --fix-types mode: deletes each type-mismatched attribute and
    recreates it with the correct SCHEMA type (Appwrite can't alter a
    type in place). Destructive per-column only; values re-sync from local.
- Align SCHEMA + Dart collectionSchema to the deployed reality for three
  fields that were double on BOTH clouds while SCHEMA said integer
  (price_adjustments.previousValue/newValue, audit_logs.amountImpact).
  Pull adapters convert double->int via _asInt/.toInt, so this is safe
  and avoids destroying audit history.

Applied to the live targets:
- Secondary: 0 type mismatches (the 3 shared fields now match SCHEMA=double).
- Primary: recreated the 8 primary-only drifted attributes
  (debts.amount->double, booking_notes.isActive->boolean,
  cash_transactions.createdBy->string, salary_cycles.{expectedAmount,
  actualPaid,remainingAmount}->integer, salary_payments.amount->integer,
  shift_notes.createdAt->integer). Verified: 0 type mismatches on both.

+++ mobile/lib/services/appwrite_sync_utils.dart
@@ -1394,8 +1394,9 @@ class AppwriteSyncUtils {
       'targetType': 'string',
       'targetUuid': 'string',
       'adjustmentType': 'string',
-      'previousValue': 'integer',
-      'newValue': 'integer',
+      // ✅ نوع مطابق للنشر الفعلي (double على الوجهتين)؛ المحوّل يحوّل double→int
+      'previousValue': 'double',
+      'newValue': 'double',
       'reason': 'string',
       'effectiveDate': 'string',
       'appliedBy': 'string',
@@ -1471,7 +1472,8 @@ class AppwriteSyncUtils {
       'timestamp': 'integer',
       'timestampIso': 'string',
       'isFinancial': 'boolean',
-      'amountImpact': 'integer',
+      // ✅ نوع مطابق للنشر الفعلي (double على الوجهتين)؛ المحوّل يحوّل double→int
+      'amountImpact': 'double',
       'action': 'string',
     },
     'payment_voids': {
